% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after_date.R
\name{after_date}
\alias{after_date}
\title{After Date Validator}
\usage{
after_date(date)
}
\arguments{
\item{date}{Reference date}
}
\value{
Validation function
}
\description{
Flags dates that are NOT after a given reference date.
}
\seealso{
Other validators: 
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{greater_than}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{matches_regex}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
