% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_columns.R
\name{compare_columns}
\alias{compare_columns}
\title{Compare Columns Validator}
\usage{
compare_columns(col1, col2, operator = "<")
}
\arguments{
\item{col1}{First column}

\item{col2}{Second column}

\item{operator}{Comparison operator.

Supported operators:
\itemize{
\item \code{"<"}
\item \code{">"}
\item \code{"<="}
\item \code{">="}
\item \code{"=="}
\item \code{"!="}
}}
}
\value{
Returns a validation function used inside
\code{\link{add_validation}}.
}
\description{
Compares values between two columns.
}
\details{
Flags rows where the comparison condition is TRUE.

Useful for:
\itemize{
\item chronological checks
\item age consistency checks
\item enrollment and exit validations
\item minimum and maximum comparisons
}
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# example 1: exit date before enrollment date ------------------------

report <- report \%>\%

  add_validation(

    label = "Exit Before Enrollment",

    columns = c(
      "date_enrolled",
      "date_exited"
    ),

    rule =
      "Exit date should not be before enrollment date",

    check = compare_columns(
      col1 = "date_exited",
      col2 = "date_enrolled",
      operator = "<"
    )
  )

# example 2: age less than minimum age -------------------------------

report <- report \%>\%

  add_validation(

    label = "Caregiver Younger Than Child",

    columns = c(
      "caregiver_age",
      "child_age"
    ),

    rule =
      "Caregiver age should be greater than child age",

    check = compare_columns(
      col1 = "caregiver_age",
      col2 = "child_age",
      operator = "<="
    )
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{validate_data}}

Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{greater_than}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{matches_regex}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
