% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ends_with.R
\name{ends_with}
\alias{ends_with}
\title{Ends With Validator}
\usage{
ends_with(text)
}
\arguments{
\item{text}{Suffix string.}
}
\value{
Returns a validation function used inside
\code{\link{add_validation}}.
}
\description{
Checks whether text values end with a given suffix.
}
\details{
Flags rows where the specified suffix is NOT found.

Useful for:
\itemize{
\item validating file extensions
\item checking email domains
\item validating identifier formats
\item enforcing naming conventions
}

The validator flags rows where:

\itemize{
\item the value does not end with the specified text
\item and the value is not missing
}
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# example 1: email should end with gmail.com -------------------------

report <- report \%>\%

  add_validation(

    label = "Invalid Gmail Address",

    columns = "email_address",

    rule =
      "Email address should end with gmail.com",

    check = ends_with("gmail.com")
  )

# example 2: files should end with .csv ------------------------------

report <- report \%>\%

  add_validation(

    label = "Invalid File Extension",

    columns = "file_name",

    rule =
      "Files should end with .csv",

    check = ends_with(".csv")
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{validate_data}}

Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{greater_than}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{matches_regex}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
