% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_to.R
\name{equal_to}
\alias{equal_to}
\title{Equal To Validator}
\usage{
equal_to(x)
}
\arguments{
\item{x}{Expected value.}
}
\value{
Returns a validation function used inside
\code{\link{add_validation}}.
}
\description{
Checks whether values are equal to an expected value.
}
\details{
Flags rows where values are NOT equal to the specified value.

Useful for:
\itemize{
\item validating fixed categories
\item checking expected status values
\item enforcing standard responses
\item validating binary indicators
}

The validator flags rows where:

\itemize{
\item the value is not equal to the expected value
\item and the value is not missing
}
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# example 1: status should be ACTIVE ---------------------------------

report <- report \%>\%

  add_validation(

    label = "Inactive Participants",

    columns = "program_status",

    rule =
      "Program status should be ACTIVE",

    check = equal_to("ACTIVE")
  )

# example 2: consent indicator should equal YES ----------------------

report <- report \%>\%

  add_validation(

    label = "Missing Consent",

    columns = "consent_provided",

    rule =
      "Consent should be marked as YES",

    check = equal_to("YES")
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{validate_data}}

Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{greater_than}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{matches_regex}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
