% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_failed.R
\name{get_failed}
\alias{get_failed}
\title{Extract Failed Records}
\usage{
get_failed(agent, validation_name)
}
\arguments{
\item{agent}{Validation object created using
\code{\link{validate_data}}.}

\item{validation_name}{Character value specifying
the validation check label.}
}
\value{
Returns a dataframe containing failed records
for the selected validation check.
}
\description{
Returns failed records for a specific validation check.
}
\details{
Useful for:
\itemize{
\item reviewing validation failures
\item exporting problematic records
\item conducting data cleaning
\item sharing error extracts with teams
}

Failed records are retrieved using the internally
stored validation ID linked to the validation label.
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# add validation -----------------------------------------------------

report <- report \%>\%

  add_validation(

    label = "Age Above 24",

    columns = "age",

    rule =
      "Participant age should not exceed 24 years",

    check = greater_than(24)
  )

# extract failed rows ------------------------------------------------

failed_records <- get_failed(
  report,
  "Age Above 24"
)

failed_records

}
\seealso{
\code{\link{add_validation}}
\code{\link{get_report}}
\code{\link{validate_data}}
}
