% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{get_report}
\alias{get_report}
\title{Generate Validation Report}
\usage{
get_report(agent)
}
\arguments{
\item{agent}{Validation object}
}
\value{
Interactive reactable report
}
\description{
Creates an interactive validation report table.
}
