% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greater_equal.R
\name{greater_equal}
\alias{greater_equal}
\title{Greater or Equal Validator}
\usage{
greater_equal(x)
}
\arguments{
\item{x}{Numeric threshold.}
}
\value{
Returns a validation function used inside
\code{\link{add_validation}}.
}
\description{
Checks whether values are greater than or equal
to a specified threshold.
}
\details{
Flags rows where values are greater than or equal
to the threshold.

Useful for:
\itemize{
\item identifying age limits
\item validating minimum thresholds
\item detecting high-risk values
\item program eligibility checks
}

The validator flags rows where:

\itemize{
\item the value is greater than or equal to
the specified threshold
\item and the value is not missing
}
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# example 1: age should be below 18 -------------------------------

report <- report \%>\%

  add_validation(

    label = "Age 18 And Above",

    columns = "age",

    rule =
      "Participant age should be below 18 years",

    check = greater_equal(18)
  )

# example 2: viral load threshold -----------------------------------

report <- report \%>\%

  add_validation(

    label = "High Viral Load",

    columns = "viral_load",

    rule =
      "Viral load should be below 1000 copies/ml",

    check = greater_equal(1000)
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{validate_data}}

Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_than}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{matches_regex}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
