% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greater_than.R
\name{greater_than}
\alias{greater_than}
\title{Greater Than Validator}
\usage{
greater_than(x)
}
\arguments{
\item{x}{numeric threshold}
}
\value{
logical expression for validation
}
\description{
Checks whether values are greater than a threshold.
}
\seealso{
Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{matches_regex}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
