% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{include}
\alias{include}
\title{Include Validator}
\usage{
include(values)
}
\arguments{
\item{values}{Character vector of allowed values.}
}
\value{
Returns a validation function used inside
\code{\link{add_validation}}.
}
\description{
Checks whether values belong to an allowed
set of categories.
}
\details{
Flags rows where values are NOT included
in the allowed list.

Useful for:
\itemize{
\item validating categorical variables
\item checking program statuses
\item enforcing standard response options
\item validating coded values
}

The validator flags rows where:

\itemize{
\item the value is not found in the allowed list
\item and the value is not missing
}
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# example 1: valid program status ------------------------------------

report <- report \%>\%

  add_validation(

    label = "Invalid Program Status",

    columns = "program_status",

    rule =
      "Program status should be ACTIVE or EXITED",

    check = include(
      c(
        "ACTIVE",
        "EXITED"
      )
    )
  )

# example 2: valid sex categories ------------------------------------

report <- report \%>\%

  add_validation(

    label = "Invalid Sex Category",

    columns = "sex",

    rule =
      "Sex should be MALE or FEMALE",

    check = include(
      c(
        "MALE",
        "FEMALE"
      )
    )
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{validate_data}}

Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{greater_than}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{matches_regex}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
