% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/less_than.R
\name{less_than}
\alias{less_than}
\title{Less Than Validator}
\usage{
less_than(x)
}
\arguments{
\item{x}{Numeric threshold.}
}
\value{
Returns a validation function used inside
\code{\link{add_validation}}.
}
\description{
Checks whether values are less than
a specified threshold.
}
\details{
Flags rows where values are less than
the threshold.

Useful for:
\itemize{
\item identifying underage participants
\item validating minimum limits
\item detecting unusually low values
\item enforcing lower boundaries
}

The validator flags rows where:

\itemize{
\item the value is less than the specified threshold
\item and the value is not missing
}
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# example 1: age should not be below 10 ------------------------------

report <- report \%>\%

  add_validation(

    label = "Age Below 10",

    columns = "age",

    rule =
      "Participant age should not be below 10 years",

    check = less_than(10)
  )

# example 2: income below minimum threshold --------------------------

report <- report \%>\%

  add_validation(

    label = "Low Household Income",

    columns = "household_income",

    rule =
      "Household income should not be below 500",

    check = less_than(500)
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{validate_data}}

Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{greater_than}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{matches_regex}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
