% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches_regex.R
\name{matches_regex}
\alias{matches_regex}
\title{Regex Validator}
\usage{
matches_regex(pattern)
}
\arguments{
\item{pattern}{Regular expression pattern.}
}
\value{
Returns a validation function used inside
\code{\link{add_validation}}.
}
\description{
Validates values against a regular expression pattern.
}
\details{
Flags rows where values do NOT match
the specified regex pattern.

Useful for:
\itemize{
\item validating phone numbers
\item checking email formats
\item validating identification numbers
\item enforcing structured text formats
}

The validator flags rows where:

\itemize{
\item the value does not match the regex pattern
\item and the value is not missing
}
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# example 1: phone number validation ---------------------------------

report <- report \%>\%

  add_validation(

    label = "Invalid Phone Number",

    columns = "phone_number",

    rule =
      "Phone number should contain 10 digits",

    check = matches_regex("^\\\\\d{10}$")
  )

# example 2: email validation ----------------------------------------

report <- report \%>\%

  add_validation(

    label = "Invalid Email Address",

    columns = "email_address",

    rule =
      "Email address should follow standard format",

    check = matches_regex(
      "^[A-Za-z0-9._\%+-]+@[A-Za-z0-9.-]+\\\\\\\\.[A-Za-z]{2,}$"
    )
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{validate_data}}

Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{greater_than}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
