% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_equal.R
\name{not_equal}
\alias{not_equal}
\title{Not Equal Validator}
\usage{
not_equal(x)
}
\arguments{
\item{x}{Value to exclude.}
}
\value{
Returns a validation function used inside
\code{\link{add_validation}}.
}
\description{
Checks whether values are equal to
a forbidden value.
}
\details{
Flags rows where values match
the excluded value.

Useful for:
\itemize{
\item excluding invalid responses
\item checking prohibited categories
\item detecting placeholder values
\item enforcing restricted entries
}

The validator flags rows where:

\itemize{
\item the value is equal to the forbidden value
\item and the value is not missing
}
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# example 1: unknown status should not exist -------------------------

report <- report \%>\%

  add_validation(

    label = "Unknown Program Status",

    columns = "program_status",

    rule =
      "Program status should not be UNKNOWN",

    check = not_equal("UNKNOWN")
  )

# example 2: age should not equal zero -------------------------------

report <- report \%>\%

  add_validation(

    label = "Zero Age Recorded",

    columns = "age",

    rule =
      "Age should not be recorded as zero",

    check = not_equal(0)
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{validate_data}}

Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{greater_than}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{matches_regex}()},
\code{\link{past_date}()},
\code{\link{range_between}()},
\code{\link{starts_with}()}
}
\concept{validators}
