% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starts_with.R
\name{starts_with}
\alias{starts_with}
\title{Starts With Validator}
\usage{
starts_with(text)
}
\arguments{
\item{text}{Prefix string.}
}
\value{
Returns a validation function used inside
\code{\link{add_validation}}.
}
\description{
Checks whether values start with
a specified prefix.
}
\details{
Flags rows where values do NOT begin
with the expected text pattern.

Useful for:
\itemize{
\item validating identification codes
\item checking program prefixes
\item enforcing naming conventions
\item validating facility identifiers
}

The validator flags rows where:

\itemize{
\item the value does not start with the prefix
\item and the value is not missing
}
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# example 1: OVC IDs should start with OVC ---------------------------

report <- report \%>\%

  add_validation(

    label = "Invalid OVC ID Prefix",

    columns = "ovc_id",

    rule =
      "OVC IDs should start with OVC",

    check = starts_with("OVC")
  )

# example 2: facility codes should start with FAC --------------------

report <- report \%>\%

  add_validation(

    label = "Invalid Facility Code Prefix",

    columns = "facility_code",

    rule =
      "Facility codes should start with FAC",

    check = starts_with("FAC")
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{validate_data}}

Other validators: 
\code{\link{after_date}()},
\code{\link{before_date}()},
\code{\link{between_dates}()},
\code{\link{compare_columns}()},
\code{\link{conditional_required}()},
\code{\link{contains}()},
\code{\link{date_before}()},
\code{\link{ends_with}()},
\code{\link{equal_to}()},
\code{\link{exclude}()},
\code{\link{future_date}()},
\code{\link{greater_equal}()},
\code{\link{greater_than}()},
\code{\link{include}()},
\code{\link{is_negative}()},
\code{\link{is_positive}()},
\code{\link{less_equal}()},
\code{\link{less_than}()},
\code{\link{matches_regex}()},
\code{\link{not_equal}()},
\code{\link{past_date}()},
\code{\link{range_between}()}
}
\concept{validators}
