% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Initialize Validation Agent}
\usage{
validate_data(data)
}
\arguments{
\item{data}{A dataframe or tibble to validate.}
}
\value{
Returns a validation agent object containing:
\itemize{
\item original dataset
\item validation results
\item failed records
}
}
\description{
Creates a validation object used to run and store
data quality validation checks.
}
\details{
The validation agent acts as the main object
for the validation workflow in the validationcheck package.

Users can:
\itemize{
\item add validation rules using \code{\link{add_validation}}
\item generate interactive reports using \code{\link{get_report}}
\item extract failed records using \code{\link{get_failed}}
}

The validation agent stores validation checks and
produces interactive data quality reports.

This function is the first step in the
validation workflow.
}
\examples{

library(validationcheck)
library(magrittr)

# initialize validation agent ----------------------------------------

report <- validate_data(sample_registry)

# add validation checks ----------------------------------------------

report <- report \%>\%

  add_validation(

    label = "Age Above 24",

    columns = "age",

    rule =
      "Participant age should not exceed 24 years",

    check = greater_than(24)
  ) \%>\%

  add_validation(

    label = "Missing Sex",

    columns = "sex",

    rule =
      "Sex should not be missing",

    check = is_missing()
  )

# generate report ----------------------------------------------------

get_report(report)

}
\seealso{
\code{\link{add_validation}}
\code{\link{get_report}}
\code{\link{get_failed}}
}
